Make Tracer an interface; TracerThread, now TracerImpl, implements it
Previously, `Tracer` and `TracerThread` were two distinct components, with the
former containing some initialization code and the latter containing the
remaining of the code.
This change moves also the initialization code to `TracerThread`, which is now
`TracerImpl`. `Tracer` is now only the interface that `TracerImpl` implements.
This is now a bit easier to understand and makes it slightly cleaner to add
methods to `Tracer`/`TracerImpl` that might be called during a capture (e.g., in
the context of
http://go/stadia-orbit-user-space-instrumentation-and-unwinding#heading=h.a3i5cfh8pexm).
The structure is also now more similar to `WindowsTracing`.

Test: Run `LinuxTracingIntegrationTests`. Take some captures on Trata.

+++ src/LinuxTracing/PerfEvent.h
@@ -234,8 +234,7 @@ inline uint64_t GetTimestamp(const PerfEvent& event) {
 }
 
 inline int GetOrderedInFileDescriptor(const PerfEvent& event) {
-  return std::visit([](const auto& arg) -> uint64_t { return arg.ordered_in_file_descriptor; },
-                    event);
+  return std::visit([](const auto& arg) -> int { return arg.ordered_in_file_descriptor; }, event);
 }
 
 }  // namespace orbit_linux_tracing
